third_party/WebKit/LayoutTests/fast/css/fontfaceset-download-error-expected.txt

CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
Test download error cases.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.fontfaces.length is 1
PASS event.fontfaces[0].family is "myfont4"
PASS event.fontfaces.length is 4
PASS failedFonts is ['myfont1', 'myfont2', 'myfont3', 'myfont5']
PASS firedEvents is ['loading', 'loadingdone', 'loadingerror']
PASS successfullyParsed is true

TEST COMPLETE
font1
font2
font3
font4
font5
